backend/ai-service/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.3
PyJWT==2.9.0
